+++ tests/kernels/quantization/test_mxfp4_triton_ep.py
@@ -29,18 +29,22 @@ class TestTritonMoeForwardExpertMap:
             torch.tensor([0, -1, 1, -1], device=device) if expert_map_present else None
         )
 
+        from vllm.utils.import_utils import import_triton_kernels
+
+        import_triton_kernels()
+
         with (
             patch("triton_kernels.topk.topk") as mock_topk,
             patch(
-                "vllm.model_executor.layers.fused_moe."
+                "vllm.model_executor.layers.fused_moe.experts."
                 "gpt_oss_triton_kernels_moe.make_routing_data"
             ) as mock_make_routing,
             patch(
-                "vllm.model_executor.layers.fused_moe."
+                "vllm.model_executor.layers.fused_moe.experts."
                 "gpt_oss_triton_kernels_moe.triton_kernel_fused_experts"
             ) as mock_fused_experts,
         ):
-            from vllm.model_executor.layers.fused_moe.gpt_oss_triton_kernels_moe import (  # noqa: E501
+            from vllm.model_executor.layers.fused_moe.experts.gpt_oss_triton_kernels_moe import (  # noqa: E501
                 triton_kernel_moe_forward,
             )
